Quickfix for resource station screens
